johnshiver_blog/requirements.txt
Django==1.8.3
Markdown==2.6.2
Pygments==2.0.2
django-authtools==1.2.0
django-model-utils==2.3.1
django-pagedown==0.1.0
djangorestframework==3.1.3
ipdb==0.8.1
ipython==3.2.1
psycopg2==2.6.1
